add blog type module and back-end api

+++ system/back-end/src/main/java/com/cy/single/blog/api/blog/LabelApi.java
@@ -4,12 +4,13 @@ import com.cy.single.blog.aspect.annotations.CheckAuth;
 import com.cy.single.blog.aspect.annotations.RecordLogger;
 import com.cy.single.blog.base.ApiResp;
 import com.cy.single.blog.base.PageResult;
-import com.cy.single.blog.pojo.req.blog.BlogLabelListReq;
-import com.cy.single.blog.pojo.req.blog.BlogLabelPageReq;
-import com.cy.single.blog.pojo.req.blog.BlogLabelReq;
+import com.cy.single.blog.pojo.req.blog.label.BlogLabelListReq;
+import com.cy.single.blog.pojo.req.blog.label.BlogLabelPageReq;
+import com.cy.single.blog.pojo.req.blog.label.BlogLabelReq;
 import com.cy.single.blog.pojo.vo.blog.BlogLabelVO;
 import com.cy.single.blog.service.BlogLabelService;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -19,10 +20,12 @@ import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
 
+import static com.cy.single.blog.utils.checkUtil.ParamValidator.checkSurrogateIds;
+
 /**
  * @Author: Lil-K
  * @Date: 2024/3/30
- * @Description: label
+ * @Description: blog label api
  */
 @Slf4j
 @RestController
@@ -51,28 +54,30 @@ public class LabelApi {
     @RecordLogger
     @CheckAuth
     @PostMapping("/save")
-    public ApiResp<String> save(@RequestBody @Validated({BlogLabelReq.GroupTypeSave.class}) BlogLabelReq req) {
+    public ApiResp<String> save(@RequestBody @Validated({BlogLabelReq.GroupLabelSave.class}) BlogLabelReq req) {
         return blogLabelService.save(req);
     }
 
     @RecordLogger
     @CheckAuth
     @PostMapping("/edit")
-    public ApiResp<String> edit(@RequestBody @Validated({BlogLabelReq.GroupTypeSave.class}) BlogLabelReq req) {
+    public ApiResp<String> edit(@RequestBody @Validated({BlogLabelReq.GroupLabelSave.class}) BlogLabelReq req) {
         return blogLabelService.edit(req);
     }
 
     @RecordLogger
     @CheckAuth
     @PostMapping("/delete")
-    public ApiResp<String> delete(@RequestBody @Validated({BlogLabelReq.GroupTypeDel.class}) BlogLabelReq req) {
+    public ApiResp<String> delete(@RequestBody @Validated({BlogLabelReq.GroupLabelDel.class}) BlogLabelReq req) {
         return blogLabelService.delete(req);
     }
 
     @RecordLogger
     @CheckAuth
     @PostMapping("/deleteBatch")
-    public ApiResp<String> deleteBatch(@RequestBody @Validated({BlogLabelReq.GroupTypeDelBatch.class}) BlogLabelReq req) {
+    public ApiResp<String> deleteBatch(@RequestBody @Validated({BlogLabelReq.GroupLabelDelBatch.class}) BlogLabelReq req) {
+        if (CollectionUtils.isEmpty(req.getSurrogateIds())) return ApiResp.failure("surrogateIds不能为空");
+        if (!checkSurrogateIds(req.getSurrogateIds())) return ApiResp.failure("surrogateIds不规范");
         return blogLabelService.deleteBatch(req);
     }
